Combined diff: by line/word/chars
diff_combined will do the diff by lines, then when it
finds different lines, it will try to make the diff by words.
If diff by words results in little number of diffs
(more than half of the line was unchanged),
then it will apply it. When comparing words, it will do the
same with characters. See specs.